Time: 23 ms (93.72%) | Memory: 48.8 MB (47.22%) - LeetSync

+++ 560-subarray-sum-equals-k/subarray-sum-equals-k.java
@@ -1,9 +1,9 @@
 class Solution {
     public int subarraySum(int[] nums, int k) {
-        HashMap<Long,Integer>map = new HashMap<>();
-        map.put(0L,1);
+        HashMap<Integer,Integer>map = new HashMap<>();
+        map.put(0,1);
         int count=0;
-        Long sum=0L;
+        int sum=0;
         for( int num:nums){
             sum += num;
             if( map.containsKey(sum-k)){
